fix file structure
