+++ src/app/shared/components/layout/navbar/navbar.component.html
@@ -1,10 +1,10 @@
-<!-- 
+
 <div id="navbar">
     <div class="navbar-inner">
         <div class="bread-crumbs"><p>{{breadcrumbs}}</p></div>
-        <div class="rightMenu">
+        <!-- <div class="rightMenu">
             <div class="searchBar"><input type="text" placeholder="Search..."></div>
             <div class="notification"></div>
-        </div>
+        </div> -->
     </div>
-</div> -->
+</div>
